+++ app/src/main/java/com/iptvapp/worker/AutoBackupWorker.kt
@@ -55,6 +55,7 @@ class AutoBackupWorker @AssistedInject constructor(
                 put("externalPlayer", prefs.externalPlayer.first())
                 put("tunneledPlaybackEnabled", prefs.tunneledPlaybackEnabled.first())
                 put("dv7FallbackEnabled", prefs.dv7FallbackEnabled.first())
+                put("audioPassthroughFallbackEnabled", prefs.audioPassthroughFallbackEnabled.first())
                 put("extraBufferingEnabled", prefs.extraBufferingEnabled.first())
                 put("silentSelfUpdateEnabled", prefs.silentSelfUpdateEnabled.first())
                 put("crashReportingEnabled", prefs.crashReportingEnabled.first())